# docs/backend/openapi.yaml
openapi: 3.0.3
info:
  title: InlineChat Gateway API
  version: 0.1.0
  description: |
    InlineChat 网关 HTTP API 草案。
    当前仅覆盖 gateway-service 暴露的 Chat/Auth/Admin 路由（不含 WebSocket 协议细节）。
servers:
  - url: http://localhost:8200
...
paths:
...
  /api/admin/v1/admin/sites:
    post:
      tags: [Admin]
      summary: 创建站点
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteRequest"
      responses:
        "201":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"

    get:
      tags: [Admin]
      summary: 查询站点列表
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/admin/v1/admin/sites/{site_id}:
    patch:
      tags: [Admin]
      summary: 更新站点信息
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: site_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSiteRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/admin/v1/admin/sites/{site_id}/status:
    patch:
      tags: [Admin]
      summary: 更新站点状态
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: site_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSiteStatusRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/admin/v1/admin/sites/{site_id}/rotate-widget-key:
    post:
      tags: [Admin]
      summary: 轮换站点 widget_key
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: site_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/admin/v1/admin/agents:
    post:
      tags: [Admin]
      summary: 创建客服账号
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAgent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"

    get:
      tags: [Admin]
      summary: 查询客服列表
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAgentList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/admin/v1/admin/agents/{id}/status:
    patch:
      tags: [Admin]
      summary: 更新客服状态
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentStatusRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAgent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/admin/v1/admin/agents/{id}/reset-password:
    post:
      tags: [Admin]
      summary: 重置客服密码
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetAgentPasswordRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAgent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/admin/v1/admin/agents/{id}/force-logout:
    post:
      tags: [Admin]
      summary: 强制客服下线
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint64
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAgent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/admin/v1/admin/audit-logs:
    get:
      tags: [Admin]
      summary: 查询审计日志
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - in: query
          name: actor_agent_id
          schema:
            type: integer
            format: uint64
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: resource_type
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      in: query
      name: limit
      description: 默认 50，最大 200
      schema:
        type: integer
        minimum: 1
        maximum: 200
    Offset:
      in: query
      name: offset
      description: 默认 0
      schema:
        type: integer
        minimum: 0
...
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 业务冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: 触发限流
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
        request_id:
          type: string
...
    CreateSiteRequest:
      type: object
      required: [site_id, name, domains]
      properties:
        site_id:
          type: string
          minLength: 4
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 128
        domains:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 3
            maxLength: 255
    UpdateSiteRequest:
      type: object
      required: [name, domains]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        domains:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 3
            maxLength: 255
    Site:
      type: object
      required: [id, site_id, name, domains, widget_key, status, created_at, updated_at]
      properties:
        id:
          type: integer
          format: uint64
        site_id:
          type: string
        name:
          type: string
        domains:
          type: array
          items:
            type: string
        widget_key:
          type: string
        status:
          type: string
          enum: [active, disabled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SiteList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Site"
    UpdateSiteStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, disabled]
    CreateAgentRequest:
      type: object
      required: [agent_id, email, password, display_name]
      properties:
        agent_id:
          type: string
          pattern: "^[0-9]{4}$"
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
          maxLength: 72
        display_name:
          type: string
          minLength: 1
          maxLength: 128
        role:
          type: string
          enum: [agent]
    AdminAgent:
      type: object
      required: [id, email, display_name, role, status, created_at, updated_at]
      properties:
        id:
          type: integer
          format: uint64
        email:
          type: string
          format: email
        display_name:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AdminAgentList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminAgent"
    UpdateAgentStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, inactive]
    ResetAgentPasswordRequest:
      type: object
      required: [new_password]
      properties:
        new_password:
          type: string
          minLength: 12
          maxLength: 72
    AuditLog:
      type: object
      required:
        [
          id,
          actor_agent_id,
          actor_email,
          actor_role,
          action,
          resource_type,
          resource_id,
          summary,
          ip,
          user_agent,
          created_at,
        ]
      properties:
        id:
          type: integer
          format: uint64
        actor_agent_id:
          type: integer
          format: uint64
        actor_email:
          type: string
        actor_role:
          type: string
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        summary:
          type: string
        ip:
          type: string
        user_agent:
          type: string
        created_at:
          type: string
          format: date-time
    AuditLogList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"